Use int type instead of int8, 16, 32...

+++ src/app/lambda-update-game/main.go
@@ -30,7 +30,7 @@ func init() {
 type UpdateGameEvent struct {
 	PlayerID string
 	GameID   string
-	Day      int16
+	Day      int
 }
 
 // UpdateGameResponse lambda response
